paper related to optimal subtokenization strategy
the paper explains a distance metric describing how different 2 clusterings are. My thinking is that it could be used to decide on what information to use and which to exclude when making clustering of amino-acid sequences present in the 2ndary structures to determine clustering via what characteristics gives us the most additional information. 

the advantage of the metric (variation of information) over similar metrics like mutual information seems to be that it is symmetrical and obeys the triangular inequality